+++ components/script/dom/testing/testbindingmaplikewithinterface.rs
